The changes to get work with long for indexes (elements, nodes, connects, etc) was finished into all NeoPZ modules. Philippe will must to be checked.

+++ Analysis/TPZNLMultGridAnalysis.h
@@ -62,7 +62,7 @@ public:
 	/** @brief Number of meshes */
 	int NMeshes() {return fMeshes.NElements();}
 	
-	TPZCompMesh *IMesh(int index);
+	TPZCompMesh *IMesh(long index);
 	
 	/**
 	 * @brief It creates a new established computational mesh in the refinement uniform
